backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "ps>=0.1.3",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
